requirements.txt
aiohttp==3.7.4.post0
arabic-reshaper==2.1.3
async-timeout==3.0.1
attrs==21.2.0
beautifulsoup4==4.10.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.7
discord==1.7.3
discord.py==1.7.3
emoji==1.6.1
future==0.18.2
gh-md-to-html==1.21.1
html5lib==1.1
idna==3.3
latex2mathml==3.61.0
Markdown==3.3.4
mdtex2html==1.2.0
multidict==5.2.0
pandoc==1.1.0
pdf2image==1.16.0
pdfkit==0.6.1
Pillow==8.4.0
plumbum==1.7.0
ply==3.11
PyPDF2==1.26.0
pypiwin32==223
python-bidi==0.4.2
python-dotenv==0.19.1
pywin32==302
reportlab==3.6.2
requests==2.26.0
shellescape==3.8.1
six==1.16.0
soupsieve==2.2.1
typing-extensions==3.10.0.2
urllib3==1.26.7
webcolors==1.11.1
webencodings==0.5.1
xhtml2pdf==0.2.5
yarl==1.7.0
